Prevent duplicate passenger emails
Removed the unique constraint on Passenger email because it was
failing before any additional actions were taken. Replaced it
with find_or_create_by using the email. This prevents duplicate emails
but the current method does not create the name of the passenger
when a new Passenger is created using the email.